add virtual in circle and add them in start on session

+++ modules/data/src/main/kotlin/agartha/data/services/IPractitionerService.kt
@@ -107,5 +107,5 @@ interface IPractitionerService : IBaseService<PractitionerDBO> {
      * @return true if practitioner successfully paid the contributionsPoints
      */
     fun payForAddingVirtualSessions(
-            practitioner: PractitionerDBO, numberOfSessions: Int): Boolean
+            practitioner: PractitionerDBO, numberOfSessions: Long): Boolean
 }
